Hotfix in delete and move task component
Fixed bug where fast-clicking 'delete document' and 'move document two days from now' would result in the document being deleted/moved more than one time. While it does not matter when moving (it does get created more than once in two days thought), deleting it more than one time will result in 404-errors after the first/initial delete request to the backend (the 404 is expected behavior, deleting a task twice or more times is not)

+++ src/components/edit/EditDialog.tsx
@@ -5,26 +5,57 @@ import EditDialogDetails from "./EditDialogDetails";
 import ArrowForwardIosOutlinedIcon from '@material-ui/icons/ArrowForwardIosOutlined';
 import Feedback from "../enums/Feedback";
 
-class EditDialog extends React.Component<Props, {}> {
+class EditDialog extends React.Component<Props, { deleteTaskDisabled: boolean, moveTaskDisabled: boolean }> {
+  constructor({props}: { props: any }) {
+    super(props);
+
+    this.state = {
+      deleteTaskDisabled: false,
+      moveTaskDisabled: false,
+    }
+  }
+
   handleClose = () => {
     const { closeDialog, updateTask } = this.props;
     updateTask();
     closeDialog();
   };
 
+  componentDidMount(): void {
+    this.setState({ deleteTaskDisabled: false });
+    this.setState({ moveTaskDisabled: false });
+  }
+
   deleteTask = async () => {
+    const { deleteTaskDisabled, moveTaskDisabled } = this.state;
     const { deleteTask } = this.props;
-    await deleteTask();
+
+    if (deleteTaskDisabled || moveTaskDisabled) {
+      return;
+    }
+
+    await this.setState({ deleteTaskDisabled: true });
+    await this.setState({ moveTaskDisabled: true });
+
+    deleteTask();
     this.handleClose();
   };
 
   moveTask = async () => {
+    const { deleteTaskDisabled, moveTaskDisabled } = this.state;
     const { currentTask, username, createNewTaskFromOldTask } = this.props;
+
+    if (deleteTaskDisabled || moveTaskDisabled) {
+      return;
+    }
+
+    await this.setState({ deleteTaskDisabled: true });
+    await this.setState({ moveTaskDisabled: true });
 
     const { day, chore, pic} = currentTask;
     const blame = username.startsWith('jan') ? 'Jan' : 'Lea';
 
-    await createNewTaskFromOldTask(currentTask.id, day, chore, pic, blame, currentTask.done);
+    createNewTaskFromOldTask(currentTask.id, day, chore, pic, blame, currentTask.done);
     this.handleClose();
   };
 
@@ -40,7 +71,7 @@ class EditDialog extends React.Component<Props, {}> {
                 <DeleteOutlineOutlinedIcon color="secondary" className="deleteIcon" />
                 <Typography className="deleteText">Aufgabe löschen</Typography>
               </div>
-              <div className="moveTask" onClick={this.moveTask}>
+              <div className="moveTask" onClick={this.moveTask} >
                 <ArrowForwardIosOutlinedIcon className="moveIcon" />
                 <Typography className="moveText">Aufgabe in zwei Tagen erneut einstellen</Typography>
               </div>
